Add fundamental variables, functions, and event listeners
The added variables are used to calculate width and height of boxes to be appended to .drawing-pad and to dictate how many boxes are to be appended on the X and Y axis.
A single div with class of "draw-box" is appended for testing reasons, and a mouseenter event listener calling paintBox() is added to above mentioned div.
The function paintBox() add "painted" to the class list of the DOM node that triggered the function to be called via above described event listener.