Set click event listener on the "Set X&Y" button with an arrow function
The arrow function assigns a number to xAxis via a prompt.
All previously written code to append divs with class names etc have been moved inside the arrow function. This is to fix issues with scope of variables, and now the boxes are appended after a number i assigned to xAxis via above mentioned promt.