Refactor paintBox() and add click event listeners on all DOM nodes with class '.draw-style-btn' to call created setDrawStyle()
Refactor paintBox() to now accomodate for different draw styles (described below). An if statement adds different inline background styles to 'this' DOM node depending on the current vaule of 'drawStyle'.
The 'shading' drawStyle adds 10% opacity. This is achieved by getting the current background style value (a rgba color) of 'this' DOM node and turns to an array to extract the alpha value and return it as a string, which in turn is converted to a floating point number and finally 0.1 is added to it.
Created setDrawStyle() gets value to the variable 'drawStyle' from the id of the DOM node that called the function through a click event listener (described below).
Click event listeners are added to all DOM nodes with class '.draw-style-btn'. Mentioned click event then calls setDrawStyle() that gives a string to the variable 'drawStyle' that the paintBox() function can use to assess what inline background style to give the DOM node that called latter mentioned function through event listeners.
Create global variables: 'drawStyleButtons', 'drawStyle', and 'alphaValue' and function scope variables, in paintBox(): 'rgba', 'rgbaValues', 'tempString', and 'tempAlpha'.
The 'alphaValue' variables value is 0 in the global scope, and is used in a template string to assign background style to all drawBox DOM nodes when they get appended.