Refactor behaviour of paintBox() if 'shading' drawStyle is selected to fix shading of colored drawBoxes
Removed the incrementation of the alpha value and instead multiply the 'r', 'g', and 'b' values by 0.9 to make the color 10% darker each time.
This fixes the issue that the 'shading' drawStyle didn't work on colored drawBoxes made by 'rainbow' drawStyle. The problem was caused by the alpha factor already beeing 1 as the colored drawBoxes were fully opaque and the 'shading' drawStyle tried to increment the alpha value that was already set to it's highest.