Add click event listeners on click of #erase-drawing
Gets rgba values of for each div of drawBoxes and subtracts 0.1 from the alpha value while leaving the 'r', 'g', and 'b' un touched.
Above is done 10 times with the use of a for loop.
After all above steps are done each div of drawBoxes gets same background rgba value as when they where first created and appended.